requirements.txt
# Core dependencies
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# API and data handling
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.4.0

# LLM integration
dspy>=0.5.7
openai>=1.3.0

# Visualization
geopandas>=0.13.0
folium>=0.14.0
streamlit-folium>=0.15.0

# Utilities
langdetect>=1.0.9
pytest>=7.3.1 